pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
package-mode = false
description = "Modular RAG pipeline with dual inference"
authors = [
    {name = "Nikhil Maheshwari"}
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "streamlit (>=1.54.0,<2.0.0)",
    "pypdf (>=6.7.3,<7.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "litellm (>=1.81.15,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "print-color (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
